Move CartDAO methods to CartDAO class
