+++ src/main/java/com/revature/p0/daos/OrderDAO.java
@@ -69,5 +69,4 @@ public class OrderDAO implements CrudDAO<Order> {
         return orders;
     }
 
-
 }
