+++ src/main/java/com/revature/p0/daos/UserDAO.java
@@ -101,51 +101,4 @@ public class UserDAO implements CrudDAO<User> {
         return creationSuccess;
     }
 
-    public boolean createCart(Cart cart) {
-        boolean isCreated = false;
-        try(Connection conn = ConnectionFactory.getInstance().getConnection()) {
-            String sql = "INSERT INTO cart (id, account_id) VALUES (?, ?)";
-
-            try(PreparedStatement ps = conn.prepareStatement(sql)) {
-                ps.setString(1, cart.getId());
-                ps.setString(2, cart.getAccountId());
-
-                isCreated = ps.executeUpdate() == 1;
-            }
-
-        } catch(SQLException e) {
-            throw new RuntimeException("Unable to connect to db", e);
-        } catch(IOException e) {
-            throw new RuntimeException("Cannot find application.properties", e);
-        } catch(ClassNotFoundException e) {
-            throw new RuntimeException("Unable to load JDBC driver", e);
-        }
-        return isCreated;
-    }
-
-    public String getCartId(String accountId) {
-
-        try(Connection conn = ConnectionFactory.getInstance().getConnection()) {
-            String sql = "SELECT * FROM cart WHERE account_id = ?";
-
-            try(PreparedStatement ps = conn.prepareStatement(sql)) {
-                ps.setString(1, accountId);
-
-                ResultSet rs = ps.executeQuery();
-
-                if (rs.next()) {
-                    return rs.getString("id");
-
-                }
-            }
-
-        } catch(SQLException e) {
-            throw new RuntimeException("Unable to connect to db", e);
-        } catch(IOException e) {
-            throw new RuntimeException("Cannot find application.properties", e);
-        } catch(ClassNotFoundException e) {
-            throw new RuntimeException("Unable to load JDBC driver", e);
-        }
-        return null;
-    }
 }
