+++ src/main/java/com/revature/p0/models/CartItems.java
@@ -32,9 +32,4 @@ public class CartItems {
         this.cartId = cartId;
     }
 
-
-
-
-
-
 }
