+++ src/main/java/com/revature/p0/screens/MainMenu.java
@@ -55,7 +55,6 @@ public class MainMenu implements IScreen {
                         System.out.println("\nPress enter to continue...");
                         scan.nextLine();
                 }
-
             }
         }
     }
