+++ src/main/java/com/revature/p0/screens/OrderScreen.java
@@ -32,6 +32,7 @@ public class OrderScreen {
 
         List<Order> allOrders = orderService.finalAllByAccountId(session.getId());
 
+        //need to add associated products
         System.out.println("past orders");
         int counter = 0;
         for(Order order : allOrders) {
